System Config admin page.

+++ flask/start-gunicorn.sh
@@ -1,6 +1,11 @@
+ps cax | grep gunicorn > /dev/null
+if [ $? -eq 0 ]; then
+  echo "Process is running."
+  exit 1
+fi
 rm -f .gunicorn.pid
 #gunicorn -w 4 -k flask_sockets.worker flaskr:app  -b '0.0.0.0:8000' --debug --log-file - --error-logfile -
-gunicorn -w 1 -k flask_sockets.worker flaskr:app  -b '0.0.0.0:8000' --debug --log-file - --error-logfile -
+gunicorn -w 1 -k flask_sockets.worker flaskr:app  -b '0.0.0.0:8000' --debug --log-file - --error-logfile -&
 pid=$!
 echo $pid > .gunicorn.pid
